661 move CA file to the root home directory
This mounts the certificates in /root/.mitmproxy as opposed to
/home/mitmproxy/.mitmproxy as mitmproxy is run by root.

Tried running it as user mitmproxy, however ~/.mitmproxy is then still
owned by root due to the fact that it is defined as VOLUME in parent
Dockerfile. This makes un unable to bypass this problem, without
mounting the volume through compose/kubernetes which will cause more
issues in the future.

In conclusion, this solution, while not great, is the best we can do
at the moment.

Signed-off-by: Szymon Łopaciuk <szymon.lopaciuk@cern.ch>